Module kfs_shell::ps2

Structs

ControlKey
A control key (ctrl, shift, alt, arrows, end, ...)

KeyEvent
A KeyEvent is the combination of a key and its state

LetterKey
A non-control key (lowercase_ascii, uppercase_ascii)

PRIMARY_PS2

PS2

Enums

Key
A key is either a letter key, a control key, or not attributed

State

Statics

PRIMARY_PS2

Functions

get_next_line

get_waitable

read_key

try_read_key