Struct kfs_shell::ps2::PS2

struct PS2 {
    status_port: Pio<u8>,
    data_port: Pio<u8>,
    event: ReadableEvent,
    is_capslocked: AtomicBool,
    is_shift: AtomicBool,
}

PS2 keyboard state.

Fields
status_port: Pio<u8>

Status Register address

data_port: Pio<u8>

Data Register address

event: ReadableEvent

IRQEvent for the PS/2 keyboard. Triggered each time the user presses a key.

is_capslocked: AtomicBool

Flips when the user has toggled the caps lock key. When set to true, all letters are returned in uppercase.

is_shift: AtomicBool

Set to true if the user is currently holding the shift key. When set to true, all letters are returned in uppercase.

Methods

impl PS2

fn handle_control_key(&self, key: ControlKey, state: State)

Handle a control key scancode

fn key_to_letter(&self, key: LetterKey) -> char

Gets the letter from the key, accounting for shift and capslock

fn read_key(&self) -> char

Waits for a single key press, and return its unicode representation.

Key presses are bufferized: if nobody is calling read_key when the user presses a key, it will be kept in a buffer until read_key is called.

fn try_read_key(&self) -> Option<char>

If a key press is pending, return its unicode representation. This can be used to implement poll-based or asynchronous reading from keyboard.

fn event_irq(&self) -> &ReadableEvent

Get a ReadableEvent for the PS2 IRQ. Waiting on this event will wait until a keypress is detected. Note that once this event is triggered, it won't trigger again until read_key or try_read_key is called.